+++ selfdrive/controls/lib/latcontrol_pid.py
@@ -2,7 +2,7 @@ import math
 
 from cereal import log
 from openpilot.selfdrive.controls.lib.latcontrol import LatControl
-from openpilot.selfdrive.controls.lib.pid import PIDController
+from openpilot.common.pid import PIDController
 
 
 class LatControlPID(LatControl):
