+++ selfdrive/controls/lib/latcontrol_torque.py
@@ -4,7 +4,7 @@ from cereal import log
 from opendbc.car.interfaces import LatControlInputs
 from openpilot.common.numpy_fast import interp
 from openpilot.selfdrive.controls.lib.latcontrol import LatControl
-from openpilot.selfdrive.controls.lib.pid import PIDController
+from openpilot.common.pid import PIDController
 from openpilot.selfdrive.controls.lib.vehicle_model import ACCELERATION_DUE_TO_GRAVITY
 
 # At higher speeds (25+mph) we can assume:
